Refactoring: moving helpers one level up to have them easier accessible
