+++ src/app/features/editor/details/components/information-link-details/information-link-details.component.ts
@@ -6,7 +6,7 @@ import { FormsModule } from '@angular/forms';
 import {KemlService} from "@app/features/editor/services/keml.service";
 import {InformationLinkType} from "@app/shared/keml/models/json/knowledge-models";
 import {Information, InformationLink, Preknowledge} from "@app/shared/keml/models/core/msg-info";
-import { InfoChoiceComponent } from '@app/shared/keml/components/helper/choices/info-choice/info-choice.component';
+import { InfoChoiceComponent } from '@app/shared/keml/components/helper/info-choice/info-choice.component';
 
 @Component({
     selector: 'app-information-link-details',
