+++ src/app/features/editor/details/components/msg-details/msg-details.component.ts
@@ -7,7 +7,7 @@ import { MatTooltip } from '@angular/material/tooltip';
 import {Information, Message, ReceiveMessage, SendMessage} from "@app/shared/keml/models/core/msg-info";
 import {ConversationPartner} from "@app/shared/keml/models/core/conversation-partner";
 import {KemlService} from "@app/features/editor/services/keml.service";
-import { InfoChoiceComponent } from '@app/shared/keml/components/helper/choices/info-choice/info-choice.component';
+import { InfoChoiceComponent } from '@app/shared/keml/components/helper/info-choice/info-choice.component';
 import { InfoInnerComponent } from '@app/shared/keml/components/helper/info-inner/info-inner.component';
 import { InputHandler} from "ngx-emfular-helper";
