+++ src/app/features/editor/editor/editor.component.ts
@@ -10,8 +10,8 @@ import { PreknowledgeComponent } from '@app/shared/keml/components/preknowledge/
 import { ConversationPartnerComponent } from '@app/shared/keml/components/cp/conversation-partner.component';
 import { AuthorComponent } from '@app/shared/keml/components/author/author.component';
 import { TextAreaSvgComponent } from "ngx-svg-graphics";
-import { DatabaseSvgComponent } from '@app/shared/keml/components/helper/svg-base-components/database-svg/database-svg.component';
-import { PersonSvgComponent } from '@app/shared/keml/components/helper/svg-base-components/person-svg/person-svg.component';
+import { DatabaseSvgComponent } from '@app/shared/keml/components/helper/database-svg/database-svg.component';
+import { PersonSvgComponent } from '@app/shared/keml/components/helper/person-svg/person-svg.component';
 import {SimulationDialogueService} from "@app/features/simulator/services/simulation-dialogue.service";
 import {ArrowMarkersComponent} from "@app/shared/keml/components/helper/arrow-markers/arrow-markers.component";
 import {KEMLIOService} from "@app/features/editor/services/keml-io.service";
