+++ src/app/shared/keml/components/helper/info-choice/info-choice.component.ts
@@ -2,9 +2,9 @@ import {Component, EventEmitter, Input, Output, TemplateRef, ViewChild} from '@a
 import {Information, Preknowledge, ReceiveMessage} from "@app/shared/keml/models/core/msg-info";
 import {MatDialog, MatDialogRef} from "@angular/material/dialog";
 import {KemlService} from "@app/features/editor/services/keml.service";
-import { PreknowledgeComponent } from '../../../preknowledge/preknowledge.component';
-import { MsgComponent } from '../../../msg/msg.component';
-import { InfoInnerComponent } from '../../info-inner/info-inner.component';
+import { PreknowledgeComponent } from '../../preknowledge/preknowledge.component';
+import { MsgComponent } from '../../msg/msg.component';
+import { InfoInnerComponent } from '../info-inner/info-inner.component';
 import { NgIf, NgFor } from '@angular/common';
 
 @Component({
