+++ src/app/shared/keml/components/helper/info-inner/info-inner.component.ts
@@ -1,7 +1,7 @@
 import {Component, Input, OnChanges, OnInit, SimpleChanges} from '@angular/core';
 import { NgIf } from '@angular/common';
 import {Information, NewInformation} from "@app/shared/keml/models/core/msg-info";
-import { DatabaseSvgComponent } from '@app/shared/keml/components/helper/svg-base-components/database-svg/database-svg.component';
+import { DatabaseSvgComponent } from '@app/shared/keml/components/helper/database-svg/database-svg.component';
 import { TextAreaSvgComponent } from "ngx-svg-graphics";
 import { IsInstrSvgComponent } from '../is-instr-svg/is-instr-svg.component';
 import {TrustComponent} from "@app/shared/keml/components/helper/trust/trust.component";
