+++ src/app/shared/keml/components/helper/is-instr-svg/is-instr-svg.component.ts
@@ -1,7 +1,7 @@
 import {Component, Input} from '@angular/core';
 import { NgIf } from '@angular/common';
 import {Information} from "@app/shared/keml/models/core/msg-info";
-import { PersonSvgComponent } from '@app/shared/keml/components/helper/svg-base-components/person-svg/person-svg.component';
+import { PersonSvgComponent } from '@app/shared/keml/components/helper/person-svg/person-svg.component';
 
 @Component({
     selector: '[is-instr]',
