+++ src/app/shared/keml/components/helper/msg-choice/msg-choice.component.ts
@@ -4,10 +4,10 @@ import {KemlService} from "@app/features/editor/services/keml.service";
 import {Message, ReceiveMessage, SendMessage} from "@app/shared/keml/models/core/msg-info";
 import {ConversationPartner} from "@app/shared/keml/models/core/conversation-partner";
 import {Author} from "@app/shared/keml/models/core/author";
-import { MsgComponent } from '../../../msg/msg.component';
-import { ConversationPartnerComponent } from '../../../cp/conversation-partner.component';
-import { AuthorComponent } from '../../../author/author.component';
-import { MsgOverviewComponent } from '../../msg-overview/msg-overview.component';
+import { MsgComponent } from '../../msg/msg.component';
+import { ConversationPartnerComponent } from '../../cp/conversation-partner.component';
+import { AuthorComponent } from '../../author/author.component';
+import { MsgOverviewComponent } from '../msg-overview/msg-overview.component';
 import { NgIf, NgFor } from '@angular/common';
 
 @Component({
