+++ src/app/shared/keml/components/helper/msg-overview/msg-overview.component.ts
@@ -1,7 +1,7 @@
 import {Component, Input} from '@angular/core';
 import {Message} from "@app/shared/keml/models/core/msg-info";
 import { TextAreaSvgComponent } from 'ngx-svg-graphics';
-import { PersonSvgComponent } from '@app/shared/keml/components/helper/svg-base-components/person-svg/person-svg.component';
+import { PersonSvgComponent } from '@app/shared/keml/components/helper/person-svg/person-svg.component';
 import { MsgInnerComponent } from '../msg-inner/msg-inner.component';
 
 @Component({
