+++ src/app/shared/keml/components/helper/person-svg/person-svg.component.ts
